# docs/swagger.yaml
basePath: /
definitions:
...
  models.CardPlayDeductDto:
    properties:
      card_no:
        type: string
      card_play_id:
        type: string
      discount_cash_amount:
        type: number
      e_bonus:
        type: integer
      e_coin:
        type: integer
      from_channel:
        type: string
    type: object
  models.CardPlayDto:
    properties:
      card_no:
        type: string
      play_branch:
        type: boolean
      play_machine:
        type: boolean
    type: object
  models.CardPlayVerifyDto:
    properties:
      card_no:
        type: string
      play_branch:
        type: string
      play_machine:
        type: string
      use_bonus:
        type: integer
      use_coin:
        type: integer
    type: object
...
  models.CardTypeDto:
    properties:
      is_active:
        type: boolean
      name:
        type: string
      show_balance:
        type: string
    type: object
...
  utils.StandardErrorResponse:
    properties:
      data:
        description: "null"
      message:
        example: Failed
        type: string
      success:
        example: false
        type: boolean
    type: object
  utils.StandardSuccessResponse:
    properties:
      data: {}
      message:
        example: Success
        type: string
      success:
        example: true
        type: boolean
    type: object
info:
  contact: {}
  description: This is a POS API with Gin and Swagger
  title: New POS APIls
  version: "1.0"
paths:
...
  /api/card-play:
    post:
      consumes:
      - application/json
      parameters:
      - description: Card Play Data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.CardPlayDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.StandardSuccessResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.StandardErrorResponse'
      security:
      - BasicAuth: []
      - BearerAuth: []
      summary: Create a new card play
      tags:
      - Card Play
  /api/card-play-deduct:
    post:
      consumes:
      - application/json
      parameters:
      - description: Card Play Deduct Data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.CardPlayDeductDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.StandardSuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.StandardErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.StandardErrorResponse'
      security:
      - BasicAuth: []
      - BearerAuth: []
      summary: Deduct card play balance
      tags:
      - Card Play
  /api/card-play-verify:
    post:
      consumes:
      - application/json
      parameters:
      - description: Card Play Verify Data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.CardPlayVerifyDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.StandardSuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.StandardErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.StandardErrorResponse'
      security:
      - BasicAuth: []
      - BearerAuth: []
      summary: Verify a card play
      tags:
      - Card Play
  /api/card-type:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.StandardSuccessResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.StandardErrorResponse'
      security:
      - BasicAuth: []
      - BearerAuth: []
      summary: Get card type all active
      tags:
      - Card Type
    post:
      consumes:
      - application/json
      parameters:
      - description: Card Type Data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.CardTypeDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.StandardSuccessResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.StandardErrorResponse'
      security:
      - BasicAuth: []
      - BearerAuth: []
      summary: Create a new card type
      tags:
      - Card Type
  /api/card-type/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.StandardSuccessResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.StandardErrorResponse'
      security:
      - BasicAuth: []
      - BearerAuth: []
      summary: Delete card type by ID
      tags:
      - Card Type
    get:
      consumes:
      - application/json
      parameters:
      - description: Id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.StandardSuccessResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.StandardErrorResponse'
      security:
      - BasicAuth: []
      - BearerAuth: []
      summary: Get card type by id
      tags:
      - Card Type
    put:
      consumes:
      - application/json
      parameters:
      - description: Card Type Data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.CardTypeDto'
      - description: Id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.StandardSuccessResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.StandardErrorResponse'
      security:
      - BasicAuth: []
      - BearerAuth: []
      summary: Update card type by ID
      tags:
      - Card Type
  /api/card-type/list:
    get:
      consumes:
      - application/json
      parameters:
      - description: Search text
        in: query
        name: search
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: skip
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.StandardSuccessResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.StandardErrorResponse'
      security:
      - BasicAuth: []
      - BearerAuth: []
      summary: Get card type list with paggination
      tags:
      - Card Type
...
  /api/crm/branch/{branchCode}:
    get:
      consumes:
      - application/json
      description: Get Branch crm by code
      parameters:
      - description: Branch Code
        in: path
        name: branchCode
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.StandardSuccessResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.StandardErrorResponse'
      security:
      - BasicAuth: []
      - BearerAuth: []
      summary: Get Branch crm by code
      tags:
      - CRM
  /api/crm/customer/{tel}:
    get:
      consumes:
      - application/json
      description: Get Customer crm by mobile number
      parameters:
      - description: Telephone Number
        in: path
        name: tel
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.StandardSuccessResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.StandardErrorResponse'
      security:
      - BasicAuth: []
      - BearerAuth: []
      summary: Get Customer crm by mobile number
      tags:
      - CRM
  /api/crm/score-type:
    get:
      consumes:
      - application/json
      description: Get score type crm
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.StandardSuccessResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.StandardErrorResponse'
      security:
      - BasicAuth: []
      - BearerAuth: []
      summary: Get score type crm
      tags:
      - CRM
...
securityDefinitions:
  BasicAuth:
    type: basic
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"